first steps towards refactoring for tests

+++ server/xpub-server/entities/manuscript/resolvers/removeSupportingFiles.js
@@ -1,33 +1,44 @@
-const { Manuscript, File, User } = require('@elifesciences/xpub-model')
+const { Manuscript, User } = require('@elifesciences/xpub-model')
 const logger = require('@pubsweet/logger')
+const { S3File } = require('@elifesciences/xpub-server')
 
-async function removeSupportingFiles(_, { id }, { user }) {
+async function removeSupportingFiles(_, id, user) {
+  return interfaceRemoveSupportingFiles(S3File, id, user)
+}
+
+async function interfaceRemoveSupportingFiles(storage, { id }, { user }) {
   const userUuid = await User.getUuidForProfile(user)
-  const manuscript = await Manuscript.find(id, userUuid)
+  let manuscript = await Manuscript.find(id, userUuid)
 
   const filesWithoutSupporting = manuscript.files.filter(
     file => file.type !== 'SUPPORTING_FILE',
   )
 
-  const files = await File.findByManuscriptId(id)
+  const files = await storage.findByManuscriptId(id)
 
-  if (files) {
+  if (files && files.length > 0) {
+    let modified = false
     files
       .filter(file => file.type === 'SUPPORTING_FILE')
       .forEach(async file => {
         try {
+          modified = true
           await file.deleteContent()
           await file.delete()
         } catch (err) {
           logger.error(`Unable to delete file ${file.id}`, err)
         }
       })
+    if (modified) {
+      manuscript.files = filesWithoutSupporting
+      manuscript = await manuscript.save()
+    }
   }
-
-  manuscript.files = filesWithoutSupporting
-  await manuscript.save()
 
   return manuscript
 }
 
-module.exports = removeSupportingFiles
+module.exports = {
+  removeSupportingFiles,
+  interfaceRemoveSupportingFiles,
+}
